requirements.txt
attrdict==2.0.0
certifi==2018.10.15
chardet==3.0.4
confusable-homoglyphs==3.2.0
cytoolz==0.9.0.1
Django==2.1.7
django-bootstrap3==11.0.0
django-jquery==3.1.0
djangorestframework==3.8.2
eth-abi==1.1.1
eth-account==0.3.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycryptodome==3.6.6
pytz==2018.5
raven==6.9.0
requests==2.20.0
rlp==1.0.3
six==1.11.0
toolz==0.9.0
urllib3==1.24.2
websockets==5.0.1
inflection==0.3.1
django-constance[database]==2.4.0
cryptography==2.6.1
git+https://gitlab.bfa.ar/adorda/django-gateway.git@v0.0.3